{{cookiecutter.repo_name}}/requirements.txt
Django>=1.7,<1.8
dj-database-url==0.3.0
django-cache-url==0.8.0
djangorestframework>=3.1.1
env-excavator>=1.3.0
